+++ reverse-proxy/balancers/round-robin.go
@@ -1,0 +1,13 @@
+package balancers
+
+import "sync"
+
+type RoundRobinBalancer struct {
+	backends []*Backend
+	mu       sync.Mutex
+	idx      int
+}
+
+func NewRoundRobinBalancer(backends []*Backend) *RoundRobinBalancer {
+	return &RoundRobinBalancer{backends: backends}
+}
